+++ app/(models)/compare/[[...slug]]/page.tsx
@@ -1,83 +1,42 @@
-'use client';
+import type { Metadata } from 'next';
+import ComparePage from '../compare-page';
 
-import { useMemo, useState } from 'react';
-import { useParams } from 'next/navigation';
-import { allModels } from '@/lib/ai/all-models';
-import type { ModelDefinition } from '@/lib/ai/all-models';
-import { Container } from '@/components/container';
-import { ModelDetails } from '@/app/(models)/models/model-details';
+const pageTitle = 'Compare Models | Sparka AI';
+const pageDescription =
+  "Compare two models from Sparka's Vercel AI Gateway catalog side-by-side. Browse all providers and models, check capabilities, context windows, and pricing.";
 
-export default function ComparePage() {
-  const params = useParams<{ slug?: string[] | string }>();
+export const metadata: Metadata = {
+  title: pageTitle,
+  description: pageDescription,
+  keywords: [
+    'Sparka',
+    'Vercel AI Gateway',
+    'compare',
+    'models',
+    'LLM',
+    'AI models',
+    'providers',
+  ],
+  openGraph: {
+    title: pageTitle,
+    description: pageDescription,
+    url: '/compare',
+    type: 'website',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: pageTitle,
+    description: pageDescription,
+  },
+  alternates: {
+    canonical: '/compare',
+  },
+  robots: {
+    index: true,
+    follow: true,
+  },
+};
 
-  const segments = useMemo(() => {
-    const raw = params?.slug;
-    if (!raw) return [] as string[];
-    if (Array.isArray(raw)) return raw;
-    return [raw];
-  }, [params]);
-
-  const [leftModelId, setLeftModelId] = useState<string | null>(() => {
-    return segments.length >= 2 ? `${segments[0]}/${segments[1]}` : null;
-  });
-
-  const [rightModelId, setRightModelId] = useState<string | null>(() => {
-    return segments.length >= 4 ? `${segments[2]}/${segments[3]}` : null;
-  });
-
-  const leftModel: ModelDefinition | null = useMemo(() => {
-    if (!leftModelId) return null;
-    return allModels.find((m) => m.id === leftModelId) || null;
-  }, [leftModelId]);
-
-  const rightModel: ModelDefinition | null = useMemo(() => {
-    if (!rightModelId) return null;
-    return allModels.find((m) => m.id === rightModelId) || null;
-  }, [rightModelId]);
-
-  function pushCompareUrl(leftId: string | null, rightId: string | null) {
-    const parts: string[] = [];
-    const pushId = (id: string) => {
-      const [a, b] = id.split('/');
-      if (a && b) parts.push(a, b);
-    };
-    if (leftId) pushId(leftId);
-    if (rightId) pushId(rightId);
-    const path = parts.length > 0 ? `/compare/${parts.join('/')}` : '/compare';
-    if (typeof window !== 'undefined') {
-      window.history.pushState({}, '', path);
-    }
-  }
-
-  function handleLeftChange(nextId: string) {
-    setLeftModelId(nextId);
-    pushCompareUrl(nextId, rightModelId);
-  }
-
-  function handleRightChange(nextId: string) {
-    setRightModelId(nextId);
-    pushCompareUrl(leftModelId, nextId);
-  }
-
-  return (
-    <Container className="p-6">
-      <div className="mb-6 flex flex-col gap-4 max-w-[450px] lg:max-w-none mx-auto">
-        <h1 className="text-2xl font-semibold ">Compare Models</h1>
-      </div>
-      <div className="grid grid-cols-1 gap-6 lg:grid-cols-2 mx-auto ">
-        <ModelDetails
-          className="mx-auto"
-          modelDefinition={leftModel}
-          onModelChangeAction={handleLeftChange}
-          enabledActions={{ goToModel: true, chat: true, compare: false }}
-        />
-        <ModelDetails
-          className="mx-auto"
-          modelDefinition={rightModel}
-          onModelChangeAction={handleRightChange}
-          enabledActions={{ goToModel: true, chat: true, compare: false }}
-        />
-      </div>
-    </Container>
-  );
+export default function Page() {
+  return <ComparePage />;
 }
