+++ app/(models)/models/page.tsx
@@ -1,9 +1,43 @@
-'use client';
-
 import { ScrollArea } from '@/components/ui/scroll-area';
 import { ModelFilters } from '@/app/(models)/models/model-filters';
 import { ModelsProvider } from '@/app/(models)/models/models-store-context';
 import { ModelsResults } from './models-results';
+import type { Metadata } from 'next';
+
+const pageTitle = 'Models | Sparka AI';
+const pageDescription =
+  'Browse models across providers form Vercel AI Gateway in Sparka AI. Filter and compare by provider, context window, and pricing.';
+
+export const metadata: Metadata = {
+  title: pageTitle,
+  description: pageDescription,
+  keywords: [
+    'Sparka',
+    'Vercel AI Gateway',
+    'models',
+    'LLM',
+    'AI models',
+    'providers',
+  ],
+  openGraph: {
+    title: pageTitle,
+    description: pageDescription,
+    url: '/models',
+    type: 'website',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: pageTitle,
+    description: pageDescription,
+  },
+  alternates: {
+    canonical: '/models',
+  },
+  robots: {
+    index: true,
+    follow: true,
+  },
+};
 
 export default function HomePage() {
   return (
